+++ src/App.js
@@ -1,0 +1,18 @@
+import "./App.css";
+import Footer from "./components/Footer";
+import Header from "./components/Header";
+import Main from "./components/Main";
+
+function App() {
+  return (
+    <>
+      <Header />
+      <hr />
+      <Main />
+      <hr />
+      <Footer />
+    </>
+  );
+}
+
+export default App;
