+++ src/components/Footer.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Footer = () => {
+  return <h1>Footer</h1>;
+};
+
+export default Footer;
