+++ src/components/Header.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Header = () => {
+  return <h1>Header</h1>;
+};
+
+export default Header;
