+++ src/components/Main.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Main = () => {
+  return <h1>Main</h1>;
+};
+
+export default Main;
